# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EqiFacture SAT CFDI Proxy API",
        "description": "Stateless proxy API for downloading CFDIs from SAT portal. No data stored.",
        "contact": {
            "name": "API Support",
            "email": "support@eqifacture.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/v1/cfdis/query": {
            "post": {
                "tags": [
                    "CFDIs (Scraping)"
                ],
                "summary": "Query CFDIs from SAT (metadata only)",
                "description": "Authenticates with FIEL and queries CFDIs. Returns metadata only, no files downloaded. Fast and synchronous.",
                "operationId": "b9db2f36d3f69a80a0251602c8530487",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "certificate",
                                    "private_key",
                                    "passphrase",
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "certificate": {
                                        "description": "FIEL certificate file (.cer) - Must be valid FIEL, not CSD",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "private_key": {
                                        "description": "FIEL private key file (.key)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "passphrase": {
                                        "description": "FIEL password",
                                        "type": "string",
                                        "example": "MySecurePassword123"
                                    },
                                    "start_date": {
                                        "description": "Start date for query range",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-01"
                                    },
                                    "end_date": {
                                        "description": "End date for query range",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-31"
                                    },
                                    "download_type": {
                                        "description": "Type of CFDIs to query",
                                        "type": "string",
                                        "default": "emitidos",
                                        "enum": [
                                            "emitidos",
                                            "recibidos",
                                            "ambos"
                                        ],
                                        "example": "emitidos"
                                    },
                                    "state_voucher": {
                                        "description": "Filter by CFDI status",
                                        "type": "string",
                                        "default": "todos",
                                        "enum": [
                                            "todos",
                                            "vigentes",
                                            "cancelados"
                                        ],
                                        "example": "todos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "CFDIs found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "count": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "cfdis": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "uuid": {
                                                                "type": "string",
                                                                "example": "b456fff2-0e87-465b-83a9-0493469cb153"
                                                            },
                                                            "rfc_emisor": {
                                                                "type": "string",
                                                                "example": "ABC123456789"
                                                            },
                                                            "nombre_emisor": {
                                                                "type": "string",
                                                                "example": "Empresa SA de CV"
                                                            },
                                                            "rfc_receptor": {
                                                                "type": "string",
                                                                "example": "XYZ987654321"
                                                            },
                                                            "total": {
                                                                "type": "string",
                                                                "example": "$4,029.98"
                                                            },
                                                            "fecha_emision": {
                                                                "type": "string",
                                                                "example": "2025-01-15T11:40:34"
                                                            },
                                                            "estado_comprobante": {
                                                                "type": "string",
                                                                "example": "Vigente"
                                                            },
                                                            "has_xml": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "has_pdf": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "FIEL authentication failed"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}